Fix a bug
# Error msg (summary)
- Input type (torch.cuda.ByteTensor)
- weight type (torch.cuda.FloatTensor)
=> should be the same

# Fix
convert input tensor to float

+++ app/loader/dataset.py
@@ -23,7 +23,7 @@ class DatasetAugPil(Dataset):
             image_np = np.array(image)
         # Apply transformations
         augmented = self.transform(image=image_np)
-        return augmented['image']
+        return augmented['image'].float()
 
 
 class DatasetAugPilLabel(Dataset):
@@ -46,4 +46,4 @@ class DatasetAugPilLabel(Dataset):
             image_np = np.array(image)
         # Apply transformations
         augmented = self.transform(image=image_np)
-        return augmented['image'], label
+        return augmented['image'].float(), label
